requirements.txt
streamlit==1.53.1
langchain==1.2.8
langchain-community==0.4.1
langchain-groq
langgraph==1.0.7
chromadb==1.4.1
sentence-transformers==5.2.2
python-docx==0.8.11
python-dotenv==1.2.1